refactor topics and qrels
