+++ pyserini/search/_base.py
