+++ pyserini/search/lucene/__init__.py
@@ -14,23 +14,15 @@
 # limitations under the License.
 #
 
-from ._base import JQuery, JDisjunctionMaxQueryGenerator, get_topics, \
-    get_topics_with_reader, get_qrels_file, get_qrels
 from ._geo_searcher import LuceneGeoSearcher
 from ._impact_searcher import JImpactSearcherResult, LuceneImpactSearcher
 from ._searcher import JLuceneSearcherResult, LuceneSimilarities, \
     LuceneFusionSearcher, LuceneSearcher
 
-__all__ = ['JDisjunctionMaxQueryGenerator',
-           'JImpactSearcherResult',
+__all__ = ['JImpactSearcherResult',
            'JLuceneSearcherResult',
-           'JQuery',
            'LuceneFusionSearcher',
            'LuceneGeoSearcher',
            'LuceneImpactSearcher',
            'LuceneSearcher',
-           'LuceneSimilarities',
-           'get_qrels',
-           'get_qrels_file',
-           'get_topics',
-           'get_topics_with_reader']
+           'LuceneSimilarities']
